backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-api-python-client==2.108.0
scikit-learn==1.3.2
numpy==1.26.0
gunicorn==21.2.0
pytest==7.4.0
httpx==0.24.1
joblib==1.3.2
scipy==1.11.3
wordninja==2.0.0
google-auth-oauthlib==1.0.0
